requirements.txt
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.42.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
